chore: optimize and clean up

+++ packages/redis-cache/src/index.ts
@@ -3,7 +3,6 @@ import type { CacheEntry, CacheStrategy } from '@dbbs/next-cache-handler-common'
 import { RedisString } from './RedisString'
 import { RedisStack } from './RedisStack'
 import { CHUNK_LIMIT } from './constants'
-import { CacheContext } from '@dbbs/next-cache-handler-common'
 
 export enum RedisAdapter {
   RedisStack,
@@ -26,33 +25,30 @@ export class RedisCache implements CacheStrategy {
     return this.redisAdapter.set(pageKey, cacheKey, data)
   }
 
-  async revalidateTag(tag: string, _ctx: CacheContext | undefined, allowCacheKeys: string[]): Promise<void> {
-    const keysToDelete = await this.redisAdapter.findByTag(tag)
-    await this.redisAdapter.client.unlink(
-      allowCacheKeys.length
-        ? keysToDelete.filter((key) => allowCacheKeys.some((allowKey) => key.endsWith(allowKey)))
-        : keysToDelete
-    )
+  async revalidateTag(tag: string, allowCacheKeys: string[]): Promise<void> {
+    const keysToDelete = await this.redisAdapter.findCacheKeys(tag, allowCacheKeys)
+    if (keysToDelete.length) await this.redisAdapter.client.unlink(keysToDelete)
   }
 
   async delete(pageKey: string, cacheKey: string): Promise<void> {
     await this.redisAdapter.client.unlink(`${pageKey}//${cacheKey}`)
   }
 
-  async deleteAllByKeyMatch(key: string, _ctx: CacheContext | undefined, allowCacheKeys: string[]): Promise<void> {
+  async deleteAllByKeyMatch(key: string, allowCacheKeys: string[]): Promise<void> {
+    if (allowCacheKeys.length) {
+      await this.redisAdapter.client.unlink(allowCacheKeys.map((allowKey) => `${key}//${allowKey}`))
+      return
+    }
+
     const keysToDelete: string[] = []
     let cursor = 0
     do {
       const result = await this.redisAdapter.client.scan(cursor, { MATCH: `${key}//*`, COUNT: CHUNK_LIMIT })
       cursor = result.cursor
-
-      if (allowCacheKeys.length) {
-        keysToDelete.push(...result.keys.filter((key) => allowCacheKeys.some((allowKey) => key.endsWith(allowKey))))
-      } else {
-        keysToDelete.push(...result.keys)
-      }
+      keysToDelete.push(...result.keys)
     } while (cursor != 0)
-    await this.redisAdapter.client.unlink(keysToDelete)
+
+    if (keysToDelete.length) await this.redisAdapter.client.unlink(keysToDelete)
     return
   }
 }
